+++ lib/twitter_with_auto_pagination/rest/api.rb
@@ -5,7 +5,11 @@ require 'twitter_with_auto_pagination/rest/timelines'
 require 'twitter_with_auto_pagination/rest/users'
 require 'twitter_with_auto_pagination/rest/uncategorized'
 
+require 'twitter_with_auto_pagination/rest/extension/clusters'
+require 'twitter_with_auto_pagination/rest/extension/favoriting'
 require 'twitter_with_auto_pagination/rest/extension/friends_and_followers'
+require 'twitter_with_auto_pagination/rest/extension/replying'
+require 'twitter_with_auto_pagination/rest/extension/unfollowing'
 
 module TwitterWithAutoPagination
   module REST
@@ -17,7 +21,11 @@ module TwitterWithAutoPagination
       include TwitterWithAutoPagination::REST::Users
       include TwitterWithAutoPagination::REST::Uncategorized
 
+      include TwitterWithAutoPagination::REST::Extension::Clusters
+      include TwitterWithAutoPagination::REST::Extension::Favoriting
       include TwitterWithAutoPagination::REST::Extension::FriendsAndFollowers
+      include TwitterWithAutoPagination::REST::Extension::Replying
+      include TwitterWithAutoPagination::REST::Extension::Unfollowing
     end
   end
 end
